% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_lda_prior.R
\name{assert_lda_prior}
\alias{assert_lda_prior}
\title{Assert a perspective topic state object}
\usage{
assert_lda_prior(x)
}
\arguments{
\item{x}{a object to assert is a perspective state object.}
}
\description{
Assert a perspective topic state object
}
\keyword{Internal}
