% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_non_zero.R
\name{assert_non_zero}
\alias{assert_non_zero}
\title{Assert a non_zero_type_topics list and a non_zero_doc_topics list}
\usage{
assert_non_zero(priors, constants, vocabulary, doc_ids)
}
\arguments{
\item{priors}{a \code{priors} object.}

\item{constants}{a \code{state_constants} object.}

\item{vocabulary}{a character vector with the vocabulary.}

\item{doc_ids}{a character vector with the document ids.}
}
\description{
Assert a non_zero_type_topics list and a non_zero_doc_topics list
}
\keyword{Internal}
