% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_state.R
\name{assert_state}
\alias{assert_state}
\title{Assert a perspective topic state object}
\usage{
assert_state(x)
}
\arguments{
\item{x}{a object to assert is a perspective state object.}
}
\description{
Assert a perspective topic state object
}
\keyword{Internal}

