% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{collapsed_sampler_cpp}
\alias{collapsed_sampler_cpp}
\title{Collapsed Gibbs sampler}
\usage{
collapsed_sampler_cpp(state, count_matrices, priors, constants)
}
\arguments{
\item{state}{a perspective model state file}

\item{count_matrices}{A list of count matrices}

\item{priors}{A list of priors}

\item{constants}{A list of constants}
}
\description{
The Standard Collapsed Gibbs sampler
}
