% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{collapsed_sampler_sa_cpp}
\alias{collapsed_sampler_sa_cpp}
\title{Collapsed Gibbs sampler using Simulated Annealing}
\usage{
collapsed_sampler_sa_cpp(state, count_matrices, priors, constants, tau)
}
\arguments{
\item{state}{a perspective model state file}

\item{count_matrices}{A list of count matrices}

\item{priors}{A list of priors}

\item{constants}{A list of constants}

\item{tau}{the annealing parameter for the iteration}
}
\description{
The Collapsed Gibbs sampler with Simulated annealing
}
