% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapsed_sampler_simulated_annealing.R
\name{collapsed_sampler_simulated_annealing}
\alias{collapsed_sampler_simulated_annealing}
\title{Collapsed Gibbs Sampler in R}
\usage{
collapsed_sampler_simulated_annealing(state, priors, params)
}
\arguments{
\item{state}{Initial state to start sampling from}

\item{priors}{Prior settings}

\item{params}{Parameters to use}
}
\description{
Collapsed Gibbs Sampler in R
}
