% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_constants.R
\name{get_constants}
\alias{get_constants}
\alias{get_lda_constants}
\title{Extract constants from a state object}
\usage{
get_constants(state)

get_lda_constants(state)
}
\arguments{
\item{state}{a \code{state} object.}
}
\value{
a \code{state_constants} object.
}
\description{
Extract constants from a state object
}
\keyword{Internal}
