% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_object.R
\name{get_priors_for_iteration}
\alias{get_priors_for_iteration}
\title{Get prior for iteration}
\usage{
get_priors_for_iteration(priors, iteration)
}
\arguments{
\item{priors}{a \code{priors} object.}

\item{iteration}{current iteration.}
}
\description{
Get prior for iteration
}
\details{
When running simulated annealing we have extract one specific prior per iteration.
}
\keyword{Internal}

