% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_count_matrices.R
\name{init_count_matrices_lda}
\alias{init_count_matrices_lda}
\title{Initialize count matrices for ordinary LDA}
\usage{
init_count_matrices_lda(state)
}
\arguments{
\item{state}{a lda state object}
}
\description{
Initialize count matrices for ordinary LDA
}
