% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_counts.R
\name{init_counts_r}
\alias{init_counts_r}
\title{Initialize count matrices (R implementation)}
\usage{
init_counts_r(state, const)
}
\arguments{
\item{state}{a state file}

\item{const}{a list with constants}
}
\description{
Initialize count matrices (R implementation)
}
\details{
Initialization of n_dk, n_vkpx, n_kpx matrices
}
\keyword{Internal}
