% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_marginal_posterior.R
\name{log_marginal_posterior}
\alias{log_marginal_posterior}
\title{Compute log marginal posterior for a state file}
\usage{
log_marginal_posterior(state, priors)
}
\arguments{
\item{state}{a \code{state} object.}

\item{priors}{a \code{priors} object.}
}
\description{
Compute log marginal posterior for a state file
}
