% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_marginal_posterior.R
\name{log_marginal_posterior_computation}
\alias{log_marginal_posterior_computation}
\title{Log-Marginal posterior}
\usage{
log_marginal_posterior_computation(count_matrices, priors)
}
\arguments{
\item{count_matrices}{a list of the following count matrices (with the same name): \code{n_dk}, \code{n_kvpx}, \code{n_kpx}, \code{n_pk}, \code{n_xk}}

\item{priors}{a \code{prior} object.}
}
\description{
Compute the log marginal posterior for the perspective topic model.
}
\keyword{Internal}
