% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_marginal_posterior_lda.R
\name{log_marginal_posterior_lda}
\alias{log_marginal_posterior_lda}
\title{Compute the log marginal posterior for an ordinary LDA model}
\usage{
log_marginal_posterior_lda(count_matrices, priors)
}
\arguments{
\item{count_matrices}{n_dk, n_k, n_vk}

\item{priors}{alpha, beta}
}
\description{
Compute the log marginal posterior for an ordinary LDA model
}
