% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{per_sampler3_cpp}
\alias{per_sampler3_cpp}
\title{C++ Cache optimized Perspective sampler with prior on Phi}
\usage{
per_sampler3_cpp(state, count_matrices, priors, constants)
}
\arguments{
\item{state}{a perspective model state file}

\item{count_matrices}{A list of count matrices}

\item{priors}{A list of priors}

\item{constants}{A list of constants}
}
\description{
C++ Cache optimized Perspective sampler with prior on Phi
}
