% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/sampler_r.R
\name{per_sampler_cpp}
\alias{per_sampler_cpp}
\alias{per_sampler_r}
\title{C++ Perspective sampler}
\usage{
per_sampler_cpp(state, count_matrices, priors, constants)

per_sampler_r(state, count_matrices, priors, constants)
}
\arguments{
\item{state}{a perspective model state file}

\item{count_matrices}{A list of count matrices}

\item{priors}{A list of priors}

\item{constants}{A list of constants}
}
\description{
C++ Perspective sampler

Perspective sampler (implemented in R as a cross-check)
}
\keyword{Internal}
