% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler_r.R
\name{per_sampler_r}
\alias{per_sampler_r}
\title{Perspective sampler (implemented in R as a cross-check)}
\usage{
per_sampler_r(state, count_matrices, priors, constants)
}
\arguments{
\item{state}{a perspective topic model state file}

\item{count}{a list of count matrices and arrays n_dk, n_vkpx, n_kpx}
}
\description{
Perspective sampler (implemented in R as a cross-check)
}
\keyword{Internal}

