% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perspective_sampler.R
\name{perspective_sampler}
\alias{perspective_sampler}
\title{Sample perspective in R}
\usage{
perspective_sampler(state, priors, params)
}
\arguments{
\item{state}{Initial state to start sampling from}

\item{priors}{Prior settings}

\item{params}{Parameters to use}
}
\description{
Sample perspective in R
}
