% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{package}
\name{PerspectiveTopicModel}
\alias{PerspectiveTopicModel}
\alias{PerspectiveTopicModel-package}
\title{Perspective Topic Model}
\description{
Package to run the perspective topic model
}

