% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pos3d}
\alias{pos3d}
\title{pos3d}
\usage{
pos3d(x, y, z, dims)
}
\arguments{
\item{x}{index 1}

\item{y}{index 2}

\item{z}{index 3}

\item{dims}{integer vector of size 3 defining the array}
}
\description{
Access positions in an vector representation of a 3D array
}

