% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_object.R
\name{prepare_prior_for_sampling}
\alias{prepare_prior_for_sampling}
\title{Prepare the prior objects for sampling}
\usage{
prepare_prior_for_sampling(priors, constants, vocabulary, doc_ids)
}
\arguments{
\item{priors}{a \code{priors} object.}

\item{constants}{a \code{state_constants} object.}

\item{vocabulary}{a character vector with the vocabulary.}

\item{doc_ids}{a character vector with the document ids.}
}
\description{
Prepare the prior objects for sampling
}
\keyword{Internal}

