% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_object.R
\name{priors}
\alias{priors}
\title{Constructor for prior object}
\usage{
priors(x = NULL, alpha = 0.1, betax0 = 0.01, betax1 = 0.01,
  alpha_pi = 0.1, beta_pi = 0.1)
}
\arguments{
\item{x}{NULL or a list with priors to set}

\item{alpha}{The alpha parameter}

\item{betax0}{The betax0 parameter}

\item{betax1}{The betax1 parameter}

\item{alpha_pi}{The alpha_pi parameter}

\item{beta_pi}{The beta_pi parameter}
}
\value{
A \code{priors} object
}
\description{
Constructor for prior object
}

