% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcategorical}
\alias{rcategorical}
\title{rcategorical}
\usage{
rcategorical(p)
}
\arguments{
\item{p}{a vector of (possoble unnormalized) probabilities to sample from}
}
\description{
Sample a categorical variable from an unnormalized proportion
}
