% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_log_marginal_posterior_table.R
\name{setup_log_marginal_posterior_table}
\alias{setup_log_marginal_posterior_table}
\title{Set up a table for when to calculate log marginal posterior}
\usage{
setup_log_marginal_posterior_table(params)
}
\arguments{
\item{params}{a \code{parameters} object.}
}
\description{
Set up a table for when to calculate log marginal posterior
}
\keyword{Internal}

