% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/throw_state_warnings.R
\name{throw_state_warnings}
\alias{throw_state_warnings}
\title{Checks state file and trow warnings for potential problems}
\usage{
throw_state_warnings(state)
}
\arguments{
\item{state}{a \code{state} object.}
}
\description{
Checks state file and trow warnings for potential problems
}
\keyword{Internal}
