% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_object.R
\name{parameters}
\alias{parameters}
\title{Constructor for parameter object}
\usage{
parameters(x = NULL, K = 10, start_iter = 2L, gibbs_iter = 10L,
  save_state_every = 10L, state_path = "perspective_",
  seed = as.integer(Sys.time()))
}
\arguments{
\item{x}{NULL or a list with priors to set}
}
\value{
A \code{parameter} object
}
\description{
Constructor for parameter object
}

